+++ src/vm/args.h
@@ -36,10 +36,10 @@
 
 #include <stdint.h>
 
-extern uint8_t client_id;
+extern uint16_t client_id;
 extern char *server_ip;
 extern int server_port;
 
-int parse_app_args(uint8_t max_ports, int argc, char *argv[]);
+int parse_app_args(uint16_t max_ports, int argc, char *argv[]);
 
 #endif /* ifndef _ARGS_H_ */
